+++ src/frontend/components/InstructorDashboardVOD/index.tsx
@@ -11,6 +11,7 @@ import { InfoWidgetModalProvider } from 'data/stores/useInfoWidgetModal';
 import { useTimedTextTrack } from 'data/stores/useTimedTextTrack';
 import { theme } from 'utils/theme/theme';
 import { InstructorDashboardVODWidgetGeneralTitle } from './widgets/InstructorDashboardVODWidgetGeneralTitle';
+import { InstructorDashboardVODWidgetUploadVideo } from './widgets/InstructorDashboardVODWidgetUploadVideo';
 
 const StyledLiveVideoInformationBarWrapper = styled(Box)`
   -webkit-box-shadow: 0px 0px 7px 5px ${normalizeColor('shadow-1', theme)};
@@ -55,6 +56,7 @@ export const InstructorDashboardVOD = () => {
       <InfoWidgetModalProvider value={null}>
         <DashboardVideoLiveWidgetsContainer>
           <InstructorDashboardVODWidgetGeneralTitle />
+          <InstructorDashboardVODWidgetUploadVideo />
         </DashboardVideoLiveWidgetsContainer>
       </InfoWidgetModalProvider>
     </Box>